+++ src/main/resources/theme/rciam-group/account/src/app/group-widgets/GroupAdminPage/GroupMembers.tsx
@@ -345,7 +345,7 @@ export const GroupMembers: FC<any> = (props) => {
               </DataListCell>,
               ...(!directMembers ? [
                 <DataListCell className="gm_vertical_center_cell" width={3} key="group-path-hd">
-                  <strong><Msg msgKey='adminGroupPathCell' /></strong>
+                  <strong><Msg msgKey='groupPath' /></strong>
                 </DataListCell>,
                 <DataListCell className="gm_vertical_center_cell" width={2} key="direct-hd">
                   <strong><Msg msgKey='adminGroupDirectMembership' /></strong>
@@ -459,7 +459,7 @@ export const GroupMembers: FC<any> = (props) => {
                       >
                         <div>
                           <Button isSmall variant="tertiary" isDisabled={member.status === "SUSPENDED"} className="gm_small_icon_button" onClick={() => {
-                            if (directMembers) { setEditMembership(member); } else { props.history.push({ pathname: '/groups/admingroups/' + member.group.id, search: '?tab=members&membership=' + member.id }) }
+                            if (member.direct) { setEditMembership(member); } else { props.history.push({ pathname: '/groups/admingroups/' + member.group.id, search: '?tab=members&membership=' + member.id }) }
                           }}>
                             <PencilAltIcon />
                           </Button>